Slurm Accounting with cost reporting
Adds in cost reporting to slurm accounting
and provides an `azslurm cost` cli command
that generates cost report of jobs and partitions,
for a given time period.

Also includes a prolog script designed to run
by slurm controller, that fetches vm size information
(such as size, region, priority, cpu and gpu counts etc)
about a job when it gets allocated, and puts it in the
admincomment field for producing cost reports.