Fix SSL certificates for Azure Maria DB
Packages a new combined SSL cert used for client auth for Azure Maria DB.
Also provides an option in autoscale.json for user to
provide a certificate URL from where the certificate
to their DB can be fetched. If certificate_url is not provided
then the packaged certificate is used.

Signed-off-by: Aditi Gaur <aditigaur@microsoft.com>