Bug Fix: Fix azslurm partitions command. (#287)
If some dynamic nodes get added on the command line without the
feature set associated with dynamic partition, then slurm still
registers those nodes but does not associate them with a partition.
In "scontrol show nodes" the partition field then appears blank
resulting in a key error. This fixes that.