Slurm topology configuration tool integration into azslurm cli (#318)
This adds a new azslurm topology CLI that can either use information from VMSS
or from the fabric manager (with elevated privileges) to infer the topology of the cluster.
It figures out the torset (top of the rack set) to understand job placement and produces a file
that can be used as a slurm topology configuration file.