change min node value to 0 (#355)
with the change to using Node count vs cores, there is no way to not use a paritition type.  Changing the `Config.MinValue = 0` for HTC, HPC and GPU will enable a value of 0 nodes in the UI, which will not build that partition.  It can also be added or removed later with azslurm scale.